Refactor modal code into reusable components and improve link behavior
